chore: shift into openapi
