+++ internal/interactivity/multiInput.go
@@ -122,7 +122,7 @@ func (m *MultiInput) HandleKeypress(key string) tea.Cmd {
 
 		return m.Focus(m.focusIndex)
 	default:
-		if len(m.inputs[m.focusIndex].AutocompleteFileExtensions) > 0 {
+		if len(m.inputs) > m.focusIndex && len(m.inputs[m.focusIndex].AutocompleteFileExtensions) > 0 {
 			if suggestions := charm_internal.SuggestionCallback(m.inputs[m.focusIndex].AutocompleteFileExtensions)(m.inputModels[m.focusIndex].Value()); len(suggestions) > 0 {
 				m.inputModels[m.focusIndex].ShowSuggestions = true
 				m.inputModels[m.focusIndex].KeyMap.AcceptSuggestion.SetEnabled(true)
